requirements.txt
asgiref==3.8.1
Django==4.2.19
django-ckeditor-5==0.2.17
django-js-asset==2.2.0
gunicorn==23.0.0
narwhals==1.28.0
networkx==3.2.1
packaging==24.2
pillow==11.1.0
plotly==6.0.0
psycopg==3.2.5
psycopg-binary==3.2.5
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
